GETP-201 refactor: auth 컴포넌트의 application 모듈 분리
