+++ src/main/java/es/princip/getp/api/controller/auth/SignUpCommandMapper.java
@@ -3,7 +3,7 @@ package es.princip.getp.api.controller.auth;
 import es.princip.getp.api.controller.CommandMapper;
 import es.princip.getp.api.controller.auth.dto.request.ServiceTermAgreementRequest;
 import es.princip.getp.api.controller.auth.dto.request.SignUpRequest;
-import es.princip.getp.domain.auth.application.command.SignUpCommand;
+import es.princip.getp.application.auth.command.SignUpCommand;
 import es.princip.getp.domain.member.model.ServiceTermAgreementData;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
