+++ src/main/java/es/princip/getp/api/security/filter/AccessTokenAuthorizationFilter.java
@@ -2,7 +2,7 @@ package es.princip.getp.api.security.filter;
 
 import es.princip.getp.api.security.exception.ExpiredTokenException;
 import es.princip.getp.api.security.exception.InvalidTokenException;
-import es.princip.getp.domain.auth.application.AccessTokenService;
+import es.princip.getp.application.auth.service.AccessTokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
