+++ src/main/java/es/princip/getp/application/auth/service/EmailVerification.java
@@ -1,6 +1,7 @@
-package es.princip.getp.domain.auth.domain;
+package es.princip.getp.application.auth.service;
 
 import lombok.AccessLevel;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
@@ -8,11 +9,11 @@ import org.springframework.data.redis.core.RedisHash;
 import org.springframework.data.redis.core.TimeToLive;
 
 import java.time.LocalDateTime;
-import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
 @Getter
 @RedisHash(value = "email_verification")
+@EqualsAndHashCode(exclude = "expiration")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class EmailVerification {
 
@@ -35,23 +36,5 @@ public class EmailVerification {
 
     public boolean verify(String verificationCode) {
         return this.verificationCode.equals(verificationCode);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof EmailVerification that)) {
-            return false;
-        }
-        return Objects.equals(email, that.email) &&
-            Objects.equals(verificationCode,that.verificationCode) &&
-            Objects.equals(createdAt, that.createdAt);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(email, verificationCode, createdAt);
     }
 }
