+++ src/main/java/es/princip/getp/application/auth/service/VerificationCodeSender.java
@@ -1,7 +1,8 @@
-package es.princip.getp.domain.auth.application;
+package es.princip.getp.application.auth.service;
 
 import es.princip.getp.domain.member.model.Email;
 
 public interface VerificationCodeSender {
+
     void send(Email email, String verificationCode);
 }
