+++ src/test/java/es/princip/getp/api/controller/auth/AuthControllerTest.java
@@ -3,7 +3,7 @@ package es.princip.getp.api.controller.auth;
 import es.princip.getp.api.controller.ControllerTest;
 import es.princip.getp.api.controller.auth.dto.request.LoginRequest;
 import es.princip.getp.api.controller.auth.dto.response.Token;
-import es.princip.getp.domain.auth.application.AuthService;
+import es.princip.getp.application.auth.service.AuthService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -28,13 +28,12 @@ class AuthControllerTest extends ControllerTest {
     @Autowired
     private AuthService authService;
 
-    @DisplayName("사용자는")
     @Nested
     class Login {
 
         final LoginRequest request = new LoginRequest(EMAIL, PASSWORD);
 
-        @DisplayName("로그인을 할 수 있다.")
+        @DisplayName("사용자는 로그인을 할 수 있다.")
         @Test
         void login() throws Exception {
             final Token token = new Token("Bearer", "${ACCESS_TOKEN}", "${REFRESH_TOKEN}");
@@ -61,11 +60,10 @@ class AuthControllerTest extends ControllerTest {
         }
     }
 
-    @DisplayName("사용자는")
     @Nested
     class ReissueAccessToken {
 
-        @DisplayName("로그인 유지를 할 수 있다.")
+        @DisplayName("사용자는 로그인 유지를 할 수 있다.")
         @Test
         void reissueAccessToken() throws Exception {
             final Token token = new Token("Bearer", "${ACCESS_TOKEN}", "${REFRESH_TOKEN}");
